combine get and get_by_index

+++ dart-ffi/src/crud.rs
@@ -9,24 +9,22 @@ use isar_core::object::isar_object::IsarObject;
 use isar_core::txn::IsarTxn;
 use serde_json::Value;
 
-struct KeySend(&'static IndexKey<'static>);
-unsafe impl Send for KeySend {}
-
-struct KeysSend(&'static [&'static IndexKey<'static>]);
-unsafe impl Send for KeysSend {}
-
 #[no_mangle]
 pub unsafe extern "C" fn isar_get(
     collection: &'static IsarCollection,
     txn: &mut IsarDartTxn,
     object: &'static mut RawObject,
-    key: Option<&'static IndexKey<'static>>,
+    key: *mut IndexKey<'static>,
 ) -> i32 {
     let object = RawObjectSend(object);
-    let key = key.map(|key| KeySend(key));
+    let key = if !key.is_null() {
+        Some(*Box::from_raw(key))
+    } else {
+        None
+    };
     isar_try_txn!(txn, move |txn| {
         let result = if let Some(key) = key {
-            collection.get_by_index(txn, key.0)?
+            collection.get_by_index(txn, &key)?
         } else {
             let oid = object.0.get_oid();
             collection.get(txn, oid)?
@@ -41,19 +39,20 @@ pub unsafe extern "C" fn isar_get_all(
     collection: &'static IsarCollection,
     txn: &mut IsarDartTxn,
     objects: &'static mut RawObjectSet,
-    keys: *const &'static IndexKey<'static>,
+    keys: *const *mut IndexKey<'static>,
 ) -> i32 {
     let objects = RawObjectSetSend(objects);
-    let keys = if keys.is_null() {
+    let keys = if !keys.is_null() {
         let slice = std::slice::from_raw_parts(keys, objects.0.get_length());
-        Some(KeysSend(slice))
+        let keys: Vec<IndexKey<'static>> = slice.iter().map(|k| *Box::from_raw(*k)).collect();
+        Some(keys)
     } else {
         None
     };
     isar_try_txn!(txn, move |txn| {
         if let Some(keys) = keys {
-            for (object, key) in objects.0.get_objects().iter_mut().zip(keys.0) {
-                let result = collection.get_by_index(txn, key)?;
+            for (object, key) in objects.0.get_objects().iter_mut().zip(keys) {
+                let result = collection.get_by_index(txn, &key)?;
                 object.set_object(result);
             }
         } else {
@@ -122,11 +121,21 @@ pub unsafe extern "C" fn isar_delete(
     collection: &'static IsarCollection,
     txn: &mut IsarDartTxn,
     oid: i64,
+    key: *mut IndexKey<'static>,
     deleted: &'static mut bool,
 ) -> i32 {
     let deleted = BoolSend(deleted);
+    let key = if !key.is_null() {
+        Some(*Box::from_raw(key))
+    } else {
+        None
+    };
     isar_try_txn!(txn, move |txn| {
-        *deleted.0 = collection.delete(txn, oid)?;
+        *deleted.0 = if let Some(key) = key {
+            collection.delete_by_index(txn, &key)?
+        } else {
+            collection.delete(txn, oid)?
+        };
         Ok(())
     })
 }
@@ -136,49 +145,31 @@ pub unsafe extern "C" fn isar_delete_all(
     collection: &'static IsarCollection,
     txn: &mut IsarDartTxn,
     oids: *const i64,
+    keys: *const *mut IndexKey<'static>,
     oids_length: u32,
     count: &'static mut u32,
 ) -> i32 {
+    let keys = if !keys.is_null() {
+        let slice = std::slice::from_raw_parts(keys, oids_length as usize);
+        let keys: Vec<IndexKey<'static>> = slice.iter().map(|k| *Box::from_raw(*k)).collect();
+        Some(keys)
+    } else {
+        None
+    };
     let oids = std::slice::from_raw_parts(oids, oids_length as usize);
     let count = UintSend(count);
     isar_try_txn!(txn, move |txn| {
-        for oid in oids {
-            if collection.delete(txn, *oid)? {
-                *count.0 += 1;
+        if let Some(keys) = keys {
+            for key in keys {
+                if collection.delete_by_index(txn, &key)? {
+                    *count.0 += 1;
+                }
             }
-        }
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn isar_delete_by_index(
-    collection: &'static IsarCollection,
-    txn: &mut IsarDartTxn,
-    key: &'static IndexKey<'static>,
-    deleted: &'static mut bool,
-) -> i32 {
-    let deleted = BoolSend(deleted);
-    isar_try_txn!(txn, move |txn| {
-        *deleted.0 = collection.delete_by_index(txn, key)?;
-        Ok(())
-    })
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn isar_delete_all_by_index(
-    collection: &'static IsarCollection,
-    txn: &mut IsarDartTxn,
-    keys: *const IndexKey<'static>,
-    keys_length: u32,
-    count: &'static mut u32,
-) -> i32 {
-    let keys = std::slice::from_raw_parts(keys, keys_length as usize);
-    let count = UintSend(count);
-    isar_try_txn!(txn, move |txn| {
-        for key in keys {
-            if collection.delete_by_index(txn, key)? {
-                *count.0 += 1;
+        } else {
+            for oid in oids {
+                if collection.delete(txn, *oid)? {
+                    *count.0 += 1;
+                }
             }
         }
         Ok(())
